+++ CHANGELOG.md
@@ -1,4 +1,14 @@
+## v2.0.0 (2024-04-26)
+
+- Add hcp_var_update module
+- Add hcp_workspace_var_update module
+- Add hcp_workspace_update module
+- Add hcp_workspace_info module
+- Rename tfc_workspaces_info to hcp_workspaces_info
+- Rename tfc_workspace_vars_info to hcp_workspace_vars_info
+- Suppress version option from v1 modules
+
 ## v1.0.0 (2024-01-12)
 
-- Create tfc_workspaces module
-- Create tfc_workspace_vars_info module
+- Add tfc_workspaces_info module
+- Add tfc_workspace_vars_info module